Fixed Panel sticking [BUG]
There was an issue with the player stopping short when clicking the stern's door panel when approaching from bellow. It was the number 1 issue with anyone I had try the game and since it's early in the playthrough, usually caused them to stop playing.

It is fixed now. The player can walk up to the panel from any angle. There was what looked to be a redundant collider on the door stopping the player short. I've removed it. With extensive testing, the player still can't walk through the door while it's closed since the door itself has a collider.

I'm guessing that someone created rough background colliders to rough up the scene. Then added the door and didn't remove the rough-in.